requirements-render.txt
Django>=4.2.0,<5.3.0
Pillow>=9.5.0
psycopg2-binary>=2.9.6
dj-database-url>=2.0.0
whitenoise>=6.5.0
gunicorn>=21.2.0
requests>=2.31.0
python-dotenv>=1.0.0
django-cors-headers>=4.2.0
django-storages>=1.14.0
boto3>=1.28.0 